Update seed to v0.8.0
