+++ frameworks/non-keyed/seed/.gitignore
@@ -2,6 +2,6 @@ node_modules
 target
 Cargo.lock
 bundled-dist/*
-!bundled-dist/js-framework-benchmark-seed.js
-!bundled-dist/js-framework-benchmark-seed_bg.wasm
+!bundled-dist/js-framework-benchmark-non-keyed-seed.js
+!bundled-dist/js-framework-benchmark-non-keyed-seed_bg.wasm
 !bundled-dist/index.html
