+++ frameworks/non-keyed/seed/Cargo.toml
@@ -8,9 +8,14 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-seed = "0.6.0"
-wasm-bindgen = "0.2.58"
-rand = {version = "0.7.3", features = ["wasm-bindgen", "small_rng"]}
+rand = "0.8.2"
+seed = "0.8.0"
+wasm-bindgen = "0.2.69"
+
+[dependencies.getrandom]
+version = "0.2.2"
+default-features = false
+features = ["js"]
 
 [profile.release]
 lto = true
