+++ frameworks/non-keyed/seed/index.html
@@ -9,8 +9,8 @@
   <body>
     <div id='main'></div>
     <script type="module">
-      import init from './js-framework-benchmark-seed.js';
-      init('./js-framework-benchmark-seed_bg.wasm');
+      import init from './js-framework-benchmark-non-keyed-seed.js';
+      init('./js-framework-benchmark-non-keyed-seed_bg.wasm');
     </script>
   </body>
 </html>
