+++ frameworks/non-keyed/seed/package.json
@@ -4,7 +4,7 @@
   "description": "Benchmark for Seed",
   "license": "ISC",
   "js-framework-benchmark": {
-    "frameworkVersion": "0.6.0"
+    "frameworkVersion": "0.8.0"
   },
   "scripts": {
     "build-dev": "rimraf bundled-dist && wasm-pack build --dev --target web --out-name js-framework-benchmark-non-keyed-seed --out-dir bundled-dist && cpr index.html bundled-dist/index.html && rimraf bundled-dist/.gitignore",
