+++ frameworks/non-keyed/seed/src/lib.rs
@@ -1,6 +1,7 @@
-use seed::{prelude::*, *};
 use rand::prelude::*;
+use seed::{prelude::*, *};
 use std::iter;
+use web_sys::Event;
 
 static ADJECTIVES: &[&'static str] = &[
     "pretty",
@@ -41,11 +42,11 @@ static NOUNS: &[&'static str] = &[
 ];
 
 // ------ ------
-// Before Mount
+//     Init
 // ------ ------
 
-fn before_mount(_: Url) -> BeforeMount {
-    BeforeMount::new().mount_point("main")
+fn init(_: Url, _: &mut impl Orders<Msg>) -> Model {
+    Model::default()
 }
 
 // ------ ------
@@ -60,8 +61,8 @@ type Position = usize;
 struct Model {
     rows: Vec<Row>,
     selected: Option<ID>,
-    generator: SmallRng,
-    previous_id: ID
+    generator: ThreadRng,
+    previous_id: ID,
 }
 
 impl Default for Model {
@@ -69,7 +70,7 @@ impl Default for Model {
         Self {
             rows: Vec::new(),
             selected: None,
-            generator: SmallRng::from_entropy(),
+            generator: thread_rng(),
             previous_id: 0,
         }
     }
@@ -81,7 +82,7 @@ struct Row {
 }
 
 impl Row {
-    fn new(previous_id: &mut ID, generator: &mut SmallRng) -> Self {
+    fn new(previous_id: &mut ID, generator: &mut ThreadRng) -> Self {
         *previous_id += 1;
         Self {
             id: *previous_id,
@@ -90,7 +91,7 @@ impl Row {
                 ADJECTIVES.choose(generator).unwrap(),
                 COLOURS.choose(generator).unwrap(),
                 NOUNS.choose(generator).unwrap(),
-            )
+            ),
         }
     }
 
@@ -119,37 +120,41 @@ fn update(msg: Msg, model: &mut Model, orders: &mut impl Orders<Msg>) {
             model.rows.clear();
             model.rows.reserve(count);
             for _ in take_units(count) {
-                model.rows.push(Row::new(&mut model.previous_id, &mut model.generator));
+                model
+                    .rows
+                    .push(Row::new(&mut model.previous_id, &mut model.generator));
             }
-        },
+        }
         Msg::Append(count) => {
             model.rows.reserve(count);
             for _ in take_units(count) {
-                model.rows.push(Row::new(&mut model.previous_id, &mut model.generator));
+                model
+                    .rows
+                    .push(Row::new(&mut model.previous_id, &mut model.generator));
             }
-        },
+        }
         Msg::Update(step) => {
             for position in (0..model.rows.len()).step_by(step) {
                 model.rows.get_mut(position).unwrap().label += " !!!";
             }
-        },
+        }
         Msg::Clear => {
             model.rows.clear();
-        },
+        }
         Msg::Swap => {
             if model.rows.len() > 998 {
                 model.rows.swap(1, 998);
             }
-        },
+        }
         Msg::Select(id) => model.selected = Some(id),
         Msg::Remove(id) => {
             model.rows.remove(Row::position(id, &model.rows));
-        },
+        }
     }
     orders.force_render_now();
 }
 
-fn take_units(count: Count) -> impl Iterator<Item=()> {
+fn take_units(count: Count) -> impl Iterator<Item = ()> {
     iter::repeat(()).take(count)
 }
 
@@ -158,24 +163,27 @@ fn take_units(count: Count) -> impl Iterator<Item=()> {
 // ------ ------
 
 fn view(model: &Model) -> Node<Msg> {
-    div![class!["container"],
+    div![
+        C!["container"],
         jumbotron(),
         table(model),
         span![
-            class!["preloadicon", "glyphicon", "glyphicon-remove"],
-            attrs!{At::from("aria-hidden") => true},
+            C!["preloadicon", "glyphicon", "glyphicon-remove"],
+            attrs! {At::from("aria-hidden") => true},
         ]
     ]
 }
 
 fn jumbotron() -> Node<Msg> {
-    div![class!["jumbotron"],
-        div![class!["row"],
-            div![class!["col-md-6"],
-                h1!["Seed"],
-            ],
-            div![class!["col-sm-6"],
-                div![class!["raw"],
+    div![
+        C!["jumbotron"],
+        div![
+            C!["row"],
+            div![C!["col-md-6"], h1!["Seed"],],
+            div![
+                C!["col-sm-6"],
+                div![
+                    C!["raw"],
                     action_button("run", "Create 1,000 rows", |_| Msg::Create(1_000)),
                     action_button("runlots", "Create 10,000 rows", |_| Msg::Create(10_000)),
                     action_button("add", "Append 1,000 rows", |_| Msg::Append(1_000)),
@@ -188,12 +196,17 @@ fn jumbotron() -> Node<Msg> {
     ]
 }
 
-fn action_button(id: &'static str, title: &'static str, on_click: impl Fn(Event) -> Msg + Clone + 'static) -> Node<Msg> {
-    div![class!["col-sm-6", "smallpad"],
+fn action_button(
+    id: &'static str,
+    title: &'static str,
+    on_click: impl FnOnce(Event) -> Msg + Clone + 'static,
+) -> Node<Msg> {
+    div![
+        C!["col-sm-6", "smallpad"],
         button![
-            class!["btn", "btn-primary", "btn-block"],
+            C!["btn", "btn-primary", "btn-block"],
             id!(id),
-            attrs!{At::Type => "button"},
+            attrs! {At::Type => "button"},
             ev(Ev::Click, on_click),
             title,
         ]
@@ -201,27 +214,40 @@ fn action_button(id: &'static str, title: &'static str, on_click: impl Fn(Event)
 }
 
 fn table(model: &Model) -> Node<Msg> {
-    table![class!["table", "table-hover", "table-striped", "test-data"],
+    table![
+        C!["table", "table-hover", "table-striped", "test-data"],
         tbody![
             id!("tbody"),
-            model.rows.iter().map(|row| view_row(row, model.selected == Some(row.id)))
+            model
+                .rows
+                .iter()
+                .map(|row| view_row(row, model.selected == Some(row.id)))
         ]
     ]
 }
 
 fn view_row(row: &Row, is_selected: bool) -> Node<Msg> {
     let id = row.id;
-    tr![attrs!{At::Class => if is_selected { AtValue::Some("danger".into()) } else { AtValue::Ignored } },
-        td![class!["col-md-1"], id.to_string()],
-        td![class!["col-md-4"], ev(Ev::Click, move |_| Msg::Select(id)),
-            a![class!["lbl"], row.label],
+    tr![
+        attrs! {At::Class => if is_selected { AtValue::Some("danger".into()) } else { AtValue::Ignored } },
+        td![C!["col-md-1"], id.to_string()],
+        td![
+            C!["col-md-4"],
+            ev(Ev::Click, move |_| Msg::Select(id)),
+            a![C!["lbl"], &row.label],
         ],
-        td![class!["col-md-1"],
-            a![class!["remove"], ev(Ev::Click, move |_| Msg::Remove(id)),
-                span![class!["glyphicon", "glyphicon-remove", "remove"], attrs!{At::from("aria-hidden") => true}],
+        td![
+            C!["col-md-1"],
+            a![
+                C!["remove"],
+                ev(Ev::Click, move |_| Msg::Remove(id)),
+                span![
+                    C!["glyphicon", "glyphicon-remove", "remove"],
+                    attrs! {At::from("aria-hidden") => true}
+                ],
             ]
         ],
-        td![class!["col-md-6"],],
+        td![C!["col-md-6"],],
     ]
 }
 
@@ -230,6 +256,6 @@ fn view_row(row: &Row, is_selected: bool) -> Node<Msg> {
 // ------ ------
 
 #[wasm_bindgen(start)]
-pub fn render() {
-    App::builder(update, view).before_mount(before_mount).build_and_start();
+pub fn start() {
+    App::start("main", init, update, view);
 }
